refactor: split SearchSelect to satisfy max-lines lint rules
Extract Dropdown to SearchSelectDropdown.tsx, pull ClearButton out as a
sub-component, and simplify handleBlur — bringing the file under 100 lines
and SearchSelect function under 70 non-blank lines.
