+++ components/shared/AddLinkModal/SearchSelect.tsx
@@ -2,6 +2,7 @@
 
 import { useRef, useState } from 'react'
 import type { EntityItem } from './AddLinkModal.types'
+import { Dropdown } from './SearchSelectDropdown'
 
 interface SearchSelectProps {
   searchQuery: string
@@ -13,39 +14,17 @@ interface SearchSelectProps {
   getEntityLabel: (item: EntityItem) => string
 }
 
-interface DropdownProps {
-  loadingSearch: boolean
-  searchResults: EntityItem[]
-  getEntityLabel: (item: EntityItem) => string
-  onSelect: (item: EntityItem) => void
-}
-
-function Dropdown({
-  loadingSearch,
-  searchResults,
-  getEntityLabel,
-  onSelect,
-}: DropdownProps) {
-  if (loadingSearch)
-    return <div className="link-modal-loading">Searching...</div>
-  if (searchResults.length === 0)
-    return <div className="link-modal-empty">No items found</div>
+function ClearButton({ onClear }: { onClear: () => void }) {
   return (
-    <ul className="link-modal-list">
-      {searchResults.slice(0, 10).map(item => (
-        <li
-          key={item.id}
-          className="link-modal-item"
-          onMouseDown={e => void e.preventDefault()}
-          onClick={() => void onSelect(item)}
-        >
-          <span className={`link-type-icon link-type-${item.type}`}>
-            {item.type === 'issue' ? '!' : 'D'}
-          </span>
-          <span className="link-modal-item-label">{getEntityLabel(item)}</span>
-        </li>
-      ))}
-    </ul>
+    <button
+      className="search-select-clear"
+      type="button"
+      aria-label="Clear selection"
+      onMouseDown={e => void e.preventDefault()}
+      onClick={onClear}
+    >
+      ×
+    </button>
   )
 }
 
@@ -68,13 +47,13 @@ export function SearchSelect({
   }
 
   function handleBlur(e: React.FocusEvent) {
-    const related = e.relatedTarget
+    const { relatedTarget } = e
     if (
-      !(related instanceof Node) ||
-      !containerRef.current?.contains(related)
-    ) {
-      setIsOpen(false)
-    }
+      relatedTarget instanceof Node &&
+      containerRef.current?.contains(relatedTarget)
+    )
+      return
+    setIsOpen(false)
   }
 
   function handleSelect(item: EntityItem) {
@@ -104,17 +83,7 @@ export function SearchSelect({
           placeholder="Search by title or number..."
           readOnly={!!selectedTarget}
         />
-        {selectedTarget && (
-          <button
-            className="search-select-clear"
-            type="button"
-            aria-label="Clear selection"
-            onMouseDown={e => void e.preventDefault()}
-            onClick={handleClear}
-          >
-            ×
-          </button>
-        )}
+        {selectedTarget && <ClearButton onClear={handleClear} />}
       </div>
       {isOpen && (
         <div className="search-select-dropdown">
